astro/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
pandas
apache-airflow-providers-postgres
apache-airflow-providers-slack==7.2.0
glob2
pytest
asyncio
flipside
sqlalchemy